# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "一个基于 Gin + GORM + Redis + RabbitMQ 的在线书店后端服务，提供用户认证、书籍管理、购物车、订单及秒杀等功能。",
        "title": "Project1_Shop API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Me",
            "email": "frottenice@outlook.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:9090",
    "basePath": "/",
    "paths": {
...
        "/admin/order/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "管理员获取所有待发货订单列表（需要管理员权限）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单管理（管理员）"
                ],
                "summary": "获取待发货订单列表（管理员）",
                "responses": {
                    "200": {
                        "description": "无权限 / 服务器繁忙",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseData"
                        }
                    }
                }
            }
        },
        "/admin/order/orderShip": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "管理员对指定订单执行发货操作（需要管理员权限）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单管理（管理员）"
                ],
                "summary": "订单发货（管理员）",
                "parameters": [
                    {
                        "description": "订单 ID",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.OrderConfirmParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误 / 订单不存在 / 无权限 / 服务器繁忙",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseData"
                        }
                    }
                }
            }
        },
...
        "/api/orderCancel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "用户取消指定未支付订单（需要登录）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单"
                ],
                "summary": "取消订单",
                "parameters": [
                    {
                        "description": "订单 ID",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.OrderConfirmParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误 / 订单不存在 / 未登录 / 服务器繁忙",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseData"
                        }
                    }
                }
            }
        },
        "/api/orderConfirm": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "用户确认收货，订单状态变更为已完成（需要登录）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单"
                ],
                "summary": "确认收货",
                "parameters": [
                    {
                        "description": "订单 ID",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.OrderConfirmParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误 / 订单不存在 / 重复操作 / 未登录 / 服务器繁忙",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseData"
                        }
                    }
                }
            }
        },
        "/api/orderCreate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "登录用户根据购物车商品创建订单，成功后自动清空购物车（需要登录）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单"
                ],
                "summary": "创建订单",
                "parameters": [
                    {
                        "description": "订单创建参数（包含商品列表）",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.OrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误 / 库存不足 / 未登录 / 服务器繁忙",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseData"
                        }
                    }
                }
            }
        },
        "/api/orderDetail/{order_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据订单 ID 获取订单详细信息（需要登录）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单"
                ],
                "summary": "获取订单详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "订单 ID",
                        "name": "order_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误 / 订单不存在 / 未登录",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseData"
                        }
                    }
                }
            }
        },
        "/api/orderPay": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "用户支付指定订单（需要登录）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单"
                ],
                "summary": "支付订单",
                "parameters": [
                    {
                        "description": "订单 ID",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.OrderConfirmParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误 / 订单不存在 / 未登录 / 服务器繁忙",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseData"
                        }
                    }
                }
            }
        },
...
        "/api/userOrders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前登录用户的所有订单及订单详情（需要登录）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单"
                ],
                "summary": "获取当前用户订单列表",
                "responses": {
                    "200": {
                        "description": "未登录 / 服务器繁忙",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseData"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.OrderConfirmParam": {
            "type": "object",
            "required": [
                "order_id"
            ],
            "properties": {
                "order_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.OrderParam": {
            "type": "object",
            "properties": {
                "book_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.OrderRequest": {
            "type": "object",
            "required": [
                "items"
            ],
            "properties": {
                "items": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/models.OrderParam"
                    }
                }
            }
        },
...
        "models.ResponseData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {}
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "格式：Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}